# Conflicts:
#	web Drinky-Winky/src/main/java/com/infoshare/dao/UserDao.java
#	web Drinky-Winky/src/main/java/com/infoshare/model/User.java
#	web Drinky-Winky/src/main/java/com/infoshare/service/UserService.java
#	web Drinky-Winky/src/main/java/com/infoshare/servlet/SearchFavouriteServlet.java
#	web Drinky-Winky/src/main/java/com/infoshare/servlet/UserViewServlet.java
#	web Drinky-Winky/src/main/java/com/infoshare/utils/Utils.java
#	web Drinky-Winky/src/main/webapp/WEB-INF/template/user-view.ftlh
#	web Drinky-Winky/src/main/webapp/WEB-INF/template/view.ftlh

+++ web Drinky-Winky/src/main/java/com/infoshare/service/DrinkService.java
@@ -10,6 +10,7 @@ import javax.ejb.EJB;
 import javax.enterprise.context.RequestScoped;
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @RequestScoped
@@ -19,14 +20,19 @@ public class DrinkService {
     @EJB
     DrinkDao drinkDao;
 
-    public void addDrink(DrinkDTO drinkDTO) {
-        Drink drink = DrinkDTO.DtoToDrink(drinkDTO);
-        drinkDao.addDrink(drink);
+    public String addDrink(DrinkDTO drinkDTO) {
+        if (drinkDao.getDrinkList().stream().noneMatch(drink -> drink.getName().equals(drinkDTO.getName()))) {
+            Drink drink = DrinkDTO.DtoToDrink(drinkDTO);
+            drinkDao.addDrink(drink);
+            return  "SUCCESS! THE DRINK HAS BEEN ADDED";
+        }
+        return "SOMETHING WENT WRONG! THE DRINK WITH THAT NAME ALREADY EXIST";
     }
 
     public void editDrink(DrinkDTO drinkDTO) {
         Drink drink = DrinkDTO.DtoToDrink(drinkDTO);
-        drinkDao.editDrink(drink); }
+        drinkDao.editDrink(drink);
+    }
 
     @Transactional
     public DrinkDTO getDrinkByName(String name) {
@@ -48,14 +54,18 @@ public class DrinkService {
 
     @Transactional
     public void deleteDrinkById(Long id) {
-        drinkDao.deleteDrinkById(id);
-        logger.info("Drink has been deleted");
+        if (drinkDao.getDrinkList().stream().anyMatch(drink -> drink.getId().equals(id))) {
+            drinkDao.deleteDrinkById(id);
+            logger.info("Drink has been deleted");
+        }
     }
 
     @Transactional
     public void deleteDrinkByName(String name) {
-        drinkDao.deleteDrinkByName(name);
-        logger.info("Drink has been deleted");
+        if (drinkDao.getDrinkList().stream().anyMatch(drink -> drink.getName().equals(name))) {
+            drinkDao.deleteDrinkByName(name);
+            logger.info("Drink has been deleted");
+        }
     }
 
     @Transactional
@@ -70,13 +80,25 @@ public class DrinkService {
     public List<DrinkDTO> getRequestDrinkList(int request, int size) {
 
         int fromIndex = (request - 1) * size;
-        int toIndex = request * size - 1;
+        int toIndex = request * size;
+
+        if (toIndex > drinkDao.getDrinkList().size()) {
+            toIndex = drinkDao.getDrinkList().size();
+        }
 
         return drinkDao.getDrinkList()
                 .stream()
                 .map(DrinkDTO::drinkToDTO)
                 .collect(Collectors.toList())
                 .subList(fromIndex, toIndex);
+    }
+
+    @Transactional
+    public Set<String> getUniqueGlassesNameList() {
+        return drinkDao.getDrinkList()
+                .stream()
+                .map(Drink::getGlassType)
+                .collect(Collectors.toSet());
     }
 
     public List<Drink> findRecipeByCategoryId(List<Long> ids) {
